backend_server/backend_server/requirements.txt
annotated-types==0.6.0
anyio==4.10.0
av==16.0.1
brotlicffi==1.0.9.2
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.3
ctranslate2==4.6.0
cycler==0.12.1
decorator==4.4.2
distro==1.9.0
easyocr==1.7.2
fastapi==0.119.0
faster-whisper==1.2.0
ffmpeg-python==0.2.0
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.9.0
future==1.0.0
google-api-core==2.26.0
google-auth==2.41.1
google-cloud-core==2.4.3
google-cloud-storage==3.4.1
google-cloud-videointelligence==2.16.2
google-cloud-vision==3.10.2
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.1
grpcio-status==1.75.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
humanfriendly==10.0
idna==3.11
imageio==2.37.0
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
jiter==0.6.1
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
langchain==0.3.27
langchain-core==0.3.77
langchain-openai==0.3.16
langchain-text-splitters==0.3.11
langsmith==0.4.32
lazy_loader==0.4
MarkupSafe==3.0.3
matplotlib==3.10.7
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.5.2
moviepy==1.0.3
mpmath==1.3.0
networkx==3.5
ninja==1.13.0
numpy==2.2.6
onnxruntime==1.23.1
openai==1.109.1
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
orjson==3.10.14
packaging==25.0
pillow==11.3.0
pip==25.2
polars==1.34.0
polars-runtime-32==1.34.0
proglog==0.1.10
proto-plus==1.26.1
protobuf==6.33.0
psutil==7.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pydub==0.25.1
pyparsing==3.2.5
pyreadline3==3.5.4
PySocks==1.7.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
scikit-image==0.25.2
scipy==1.16.2
setuptools==80.9.0
shapely==2.1.2
six==1.17.0
sniffio==1.3.0
SQLAlchemy==2.0.44
starlette==0.48.0
sympy==1.14.0
tenacity==9.1.2
tifffile==2025.10.16
tiktoken==0.9.0
tokenizers==0.22.1
torchvision==0.24.0
tqdm==4.67.1
typing_extensions==4.15.0
typing-inspection==0.4.2
ultralytics-thop==2.0.17
urllib3==2.5.0
uvicorn==0.38.0
wheel==0.45.1
win-inet-pton==1.1.0
zstandard==0.24.0
